refactor: rename route from "/blog/create" to "/blog/management"
